Second commit
Added a second condition to the reverse-general function: if the input is a list, use cdr to remove all but the first value and reverse it by recursively calling the reverse-general function. Do the same with the elements at the beginning of the list and recombine them using the list function. Finally, use the append function to combine the individual lists.