Problem 3 commit 9
Test whether or not the value in the list is equal to the letter z. If it is recursively call the function with the remainder of the list as the parameter. If the value is not a letter and is a number, add that value to the remainder of the list that is used as a parameter to the recursively called function. Return the value of the sum.