Problem 4 commit 2
End of min-above-max function. If the second list is null, display the minimum of the first list by calling the min-list function (to be detailed in the next commit). Otherwise, call the second-smallest function (to be detailed in later commit) with the minimum of the second list and the entirety of the first list as parameters.