Problem 4 commit 4
Continuation of min-list function. If the list only consists of one value (verified by checking if the remainder of the list is null using cdr), return the one value in the list. Otherwise, compare the first value in the list with the remainder of the list by recursively calling the function. When the minimum is found, return it. If the first value is not smaller, recursively call the function with the remainder of the list.