Problem 4 commit 5
second-smallest function added. This function takes list 1 and the minimum value of list 2 as parameters. If the first value in list 1 is equal to the minimum of list 2, the minimum is removed from list 1 by recursively calling the function with the remainder of the list. The values that remain are put into a new list which is then sent to the minimum function to find its minimum. The minimum that is produced is the smallest value in list 1 that is larger than the minimum of list 2.